requirements.txt
# Reddit Sentiment Analysis - Dependencies
# Install with: pip install -r requirements.txt

# Data Collection & Processing
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0

# Sentiment Analysis - Lexicon Based
nltk>=3.8.1
textblob>=0.17.1

# Sentiment Analysis - Transformer Based
transformers>=4.35.0
torch>=2.0.0

# Sentiment Analysis - LLM (Azure OpenAI)
openai>=1.3.0

# Evaluation & Visualization
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Environment Variables
python-dotenv>=1.0.0

# Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.25.0

